Refactor DirectiveController and UserService
This commit refactors the DirectiveController and UserService classes. It introduces a new UserService dependency in the DirectiveController constructor and updates the store method to use the UserService to create a new Directive and associated User. It also updates the update method to extract and update the User attributes separately from the Directive attributes. The UserService class is also updated to extract and remove the User attributes before creating the User.